requirements.txt
Flask==3.0.3
Werkzeug==3.0.4
line-bot-sdk==2.0.1
openai>=1.0.0